modified HotelDAO and the interface, created a factory class to inject dependency.

+++ src/hotellab/DB_Accessor.java
@@ -19,8 +19,8 @@ public abstract List<Map<String, Object>> getRecords(String table) throws SQLExc
 
 public abstract int insertRecord(String table, List<String> colNames, List values);
 
-public abstract int updateRecord(String table, String primaryKey, int pk, String colName, Object value)throws SQLException, Exception;
+public abstract int updateRecord(String table, String primaryKey, Long pk, String colName, Object value)throws SQLException, Exception;
 
-public abstract int deleteRecord(String table, String primaryKey, int pk)throws SQLException, Exception;
+public abstract int deleteRecord(String table, String primaryKey, Long pk)throws SQLException, Exception;
     
 }
